Fix implicit 64-bit to 32-bit value conversion warnings when using the gcc(1) -Wshorten-64-to-32 warning option.

+++ osh.c
@@ -1128,7 +1128,7 @@ vacount(const char **vp)
 	for (p = vp; *p != NULL; p++)
 		;	/* nothing */
 
-	return p - vp;
+	return (int)(p - vp);
 }
 
 /*
